requirements.txt
-e ./package # awesome-streamlit extensions

streamlit
langchain
openai
python-dotenv
fastapi
uvicorn
tiktoken
bs4
pydantic
streamlit_chat
tinydb
fpdf